updated work for custom hook
